python_client/requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0
httpx>=0.26.0
openai>=1.10.0
python-dotenv>=1.0.0
